core/requirements.txt
PyYAML
six
jinja2
pyauto.util>=0.0.3a3
